healthcare_rag_backend/requirements.txt
# FastAPI server
fastapi==0.115.4
uvicorn[standard]==0.35.0
python-dotenv==1.0.1

# OpenAI API
openai==1.58.1

# Redis
redis==5.2.0

# Utilities
pydantic==2.10.2
httpx==0.27.0
streamlit==1.53.0
python-multipart==0.0.17
